Update sections
Sections now properly update dimensions when changing text within to match new length of text. Also, section properly updates to move when a line is made or removed which would move it. Sections can also be clicked and dragged to move to position in the text. Still some bugs to fix and polish and is missing subsections and hierarchy function in this version