Update Section hierarchy
Section hierarchy displays the sections in the main text page. Parent sections display as a dropdown with their own name and any children's name as a label within. Still working on clicking on a section label to go to location and updating the hierarchy UI in real time when sections are created